// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/questionType.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type QuestionType = 0 | 1 | 2 | 3 | 4 | 5 | 6;

export const QuestionType = {
    NUMBER_0: 0 as QuestionType,
    NUMBER_1: 1 as QuestionType,
    NUMBER_2: 2 as QuestionType,
    NUMBER_3: 3 as QuestionType,
    NUMBER_4: 4 as QuestionType,
    NUMBER_5: 5 as QuestionType,
    NUMBER_6: 6 as QuestionType
};